requirements.txt
APScheduler==3.6.1
Click==7.0
Flask==1.1.1
Flask-API==1.1
Flask-Cors==3.0.8
Flask-DebugTool==0.1.8
Flask-HTTPAuth==3.3.0
Flask-SQLAlchemy==2.4.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Naked==0.1.31
PyMySQL==0.9.3
PyYAML==5.1.2
SQLAlchemy==1.3.6
Werkzeug==0.15.5
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
flask-sqlacodegen==1.1.6.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
inflect==2.1.0
itsdangerous==1.1.0
pip==19.0.3
pycparser==2.19
pycryptodome==3.8.2
pymssql==2.1.4
pytz==2019.2
pywin32==224
redis==3.3.8
requests==2.22.0
setuptools==40.8.0
shellescape==3.4.1
six==1.12.0
sqlacodegen==2.1.0
tzlocal==2.0.0
urllib3==1.25.3
winpexpect==1.5




